Use Pandoc for requirements/guidelines (#18)
The usage of Pandoc must be explicitly enabled by setting `CABF_PANDOC_ENABLED` environment variable to `true`, to avoid failure when Pandoc is not installed on the system.

If this environment variable is not set Markdown will be used as fallback.

Pandoc rendering will only be used for the requirement documents as they use Pandoc specific markup. The rest of the website will continue the goldmark markdown rendering for performance and better support in Hugo.